# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  responses.CommonResponse:
    properties:
      data: {}
      errors: {}
      request_id:
        type: string
      response_code:
        type: integer
      response_message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive RESTful API built with Go, Fiber, and PostgreSQL using
    Hexagonal Architecture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go RESTful API - Hexagonal Architecture
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of roles with optional filtering
      parameters:
      - default: 10
        description: 'Number of roles to return (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 'Number of roles to skip (default: 0)'
        in: query
        minimum: 0
        name: offset
        type: integer
      - description: Filter by active status (true/false)
        enum:
        - "true"
        - "false"
        in: query
        name: active
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "400":
          description: Bad request - Invalid parameters
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role with the provided information
      parameters:
      - description: Role creation request
        in: body
        name: role
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "409":
          description: Conflict - Role with same slug already exists
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role (soft delete)
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "400":
          description: Bad request - Invalid role ID
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Role does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Retrieve a specific role by its unique identifier
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role details
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid role ID
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Role does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update an existing role's information
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Role update request
        in: body
        name: role
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Role does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - roles
  /roles/{id}/activate:
    put:
      consumes:
      - application/json
      description: Activate a deactivated role
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role activated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid role ID
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Role does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Activate role
      tags:
      - roles
  /roles/{id}/deactivate:
    put:
      consumes:
      - application/json
      description: Deactivate an active role
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deactivated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Invalid role ID
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Role does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Deactivate role
      tags:
      - roles
  /roles/search:
    get:
      consumes:
      - application/json
      description: Search roles by name or description
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - default: 10
        description: 'Number of roles to return (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 'Number of roles to skip (default: 0)'
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of matching roles
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "400":
          description: Bad request - Invalid parameters
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Search roles
      tags:
      - roles
  /roles/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Retrieve a role by its URL-friendly slug
      parameters:
      - description: Role slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role details
          schema:
            allOf:
            - $ref: '#/definitions/responses.CommonResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request - Slug is required
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "404":
          description: Not found - Role does not exist
          schema:
            $ref: '#/definitions/responses.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get role by slug
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"